feat(ios): add basics of ios detection app
This is pretty much taken straight from
https://medium.freecodecamp.org/ios-coreml-vision-image-recognition-3619
cf319d0b with uses Apple's CoreML/Vision APIs, I'll just use this as a
basis and add in tensorflow-lite instead.

This doesn't look totally trivial according to
https://stackoverflow.com/questions/47355595/tensorflow-lite-swift-setup
-bridge-header-includes-root-folder as tensorflow lite has an
Objective-C interface not a Swift one.

I actually had already played around with adding TF-Lite into the mix
so there might be some remenance of that in this commit that wasn't
intentional.