feat(tflite): add the beginnings of the use of tflite
This doesn't actually use any TF lite functionality yet but demonstrates
how to include Objective-C++ relying on stl libraries in Swift.

Here we define our wrapper class `ImageRecogniser` that will hold the
model and perform the request to get the most likely label, given an
image as input. I'm basing my work on
https://github.com/tensorflow/tensorflow/blob/master/tensorflow/contrib/
lite/examples/ios/camera/CameraExampleViewController.h

One complication is introduced in that I want to be able to use some of
the STL structures that are used in the above like vector<> which I'd
quite like to use. Not strictly necessary but I thought it would be a
nice thing to try out, given that I'm more used to using these than
anything else. The approach to be able to do this, which I found on
https://stackoverflow.com/questions/36948995/how-to-make-objective-c-cla
ss-using-stdvector-made-available-to-swift-classes resolves the issue of
Swift bridge not having the STL in its include path by simply hiding
this implementation detail from it, and only using this within
Objective-C++ compilation.