chore(keras): install keras into pipenv
Installs keras, and a tensorflow with AVX2 FMA CPU instruction support.
The default tensorflow from pypi doesn't come with these enabled, taking
cue from
https://stackoverflow.com/questions/47068709/your-cpu-supports-instructi
ons-that-this-tensorflow-binary-was-not-compiled-to-u and a desire to
ensure that we are building models as fast as possible I'm installing a
version that does have support.

Note to do this I have also had to update python version to 3.7, and
this will also mean that we are now osx specific.